split test getcmd
